Rename SessionStub to SessionMock and add a real SessionStub
